+++ state/fire.go
@@ -622,7 +622,7 @@ func (i *Instance[S, E, C]) runActions(refs []Ref, entity C, tr *Trace) (effects
 			return effects, a.Name, aerr
 		}
 		effects = append(effects, e)
-		tr.EffectsEmitted = append(tr.EffectsEmitted, fmt.Sprintf("%s:%s", a.Name, typeName(e)))
+		tr.EffectsEmitted = append(tr.EffectsEmitted, fmt.Sprintf("%s:%s", a.Name, effectLabel(e)))
 		if ms, ok := commMicrostep(e); ok {
 			tr.Microsteps = append(tr.Microsteps, ms)
 		}